# internal/models/profileData.go
package models

// Структура данных профилей, содержащихся в БД, для хранения и вывода
type ProfileData struct {
	Login     string `json:"login"`     // логин профиля, используется при авторизации, является первичным ключом для БД, должен быть уникальным
	FirstName string `json:"firstName"` // имя пользователя
	LastName  string `json:"lastName"`  // фамилия пользователя
}

// Структура данных, содержащая только логин
type LoginData struct {
	Login string `json:"login"` // логин профиля, используется при авторизации, является первичным ключом для БД, должен быть уникальным
}

// Структура данных профилей, включающая все данные пользователя, включая пароль
type FullProfileData struct {
	Login     string `json:"login"`     // логин профиля, используется при авторизации, является первичным ключом для БД, должен быть уникальным
	FirstName string `json:"firstName"` // имя пользователя
	LastName  string `json:"lastName"`  // фамилия пользователя
	Password  string `json:"password"`  // пароль для входа нового пользователя (должен быть не длиннее 72 символов)
}

// Структура данных, содержащая пару логин-пароль, используется для смены пароля пользователя
type NewPasswordForProfile struct {
	Login       string `json:"login"`       // логин профиля, используется при авторизации, является первичным ключом для БД, должен быть уникальным
	NewPassword string `json:"newPassword"` // новый пароль для входа пользователя (должен быть не длиннее 72 символов)
}